Rewrite prepareRefactorings method of RefactoringUtil.
