+++ src/main/java/org/jetbrains/research/groups/ml_methods/refactoring/RefactorIntentionAction.java
@@ -10,19 +10,18 @@ import com.intellij.psi.PsiFile;
 import com.intellij.util.IncorrectOperationException;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.research.groups.ml_methods.algorithm.refactoring.CalculatedRefactoring;
-import org.jetbrains.research.groups.ml_methods.algorithm.refactoring.Refactoring;
+import org.jetbrains.research.groups.ml_methods.algorithm.refactoring.MoveToClassRefactoring;
 import org.jetbrains.research.groups.ml_methods.utils.RefactoringUtil;
 
 import java.util.Collections;
 
 public class RefactorIntentionAction extends BaseIntentionAction {
     private final AnalysisScope scope;
-    private final Refactoring refactoring;
+    private final MoveToClassRefactoring refactoring;
 
     RefactorIntentionAction(String unit, String to, AnalysisScope scope) {
         this.scope = scope;
-        this.refactoring = Refactoring.createRefactoring(unit, to, false, scope);
+        this.refactoring = MoveToClassRefactoring.createRefactoring(unit, to, false, scope);
     }
 
     @NotNull
