+++ src/main/java/org/jetbrains/research/groups/ml_methods/refactoring/RefactoringSessionInfo.java
@@ -1,7 +1,7 @@
 package org.jetbrains.research.groups.ml_methods.refactoring;
 
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.research.groups.ml_methods.algorithm.refactoring.Refactoring;
+import org.jetbrains.research.groups.ml_methods.algorithm.refactoring.MoveToClassRefactoring;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -11,7 +11,7 @@ import java.util.stream.Collectors;
  * Objects of this class contain information about one user interaction with the plugin.
  * Refactorings are suggested to user and he accepts some of them and rejects the other.
  * Information about what refactorings were accepted and what refactorings were rejected is
- * stored in objects of this class. Each {@link Refactoring} is represented by
+ * stored in objects of this class. Each {@link MoveToClassRefactoring} is represented by
  * {@link RefactoringFeatures} derived from it. All data from this class is intended to be written
  * to log for further analysis.
  */
@@ -27,8 +27,8 @@ public final class RefactoringSessionInfo {
      * @param rejectedRefactorings refactorings that were rejected.
      */
     public RefactoringSessionInfo(
-        final @NotNull List<Refactoring> acceptedRefactorings,
-        final @NotNull List<Refactoring> rejectedRefactorings
+        final @NotNull List<MoveToClassRefactoring> acceptedRefactorings,
+        final @NotNull List<MoveToClassRefactoring> rejectedRefactorings
     ) {
         acceptedRefactoringsFeatures =
             acceptedRefactorings.stream()
