+++ src/main/java/org/jetbrains/research/groups/ml_methods/ui/ClassRefactoringPanel.java
@@ -7,7 +7,7 @@ import com.intellij.ui.components.JBPanel;
 import com.intellij.ui.table.JBTable;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.research.groups.ml_methods.algorithm.refactoring.CalculatedRefactoring;
-import org.jetbrains.research.groups.ml_methods.algorithm.refactoring.Refactoring;
+import org.jetbrains.research.groups.ml_methods.algorithm.refactoring.MoveToClassRefactoring;
 import org.jetbrains.research.groups.ml_methods.config.RefactoringPreferencesLog;
 import org.jetbrains.research.groups.ml_methods.refactoring.RefactoringSessionInfo;
 import org.jetbrains.research.groups.ml_methods.refactoring.RefactoringSessionInfoRenderer;
@@ -22,7 +22,6 @@ import java.awt.*;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Predicate;
@@ -160,8 +159,8 @@ class ClassRefactoringPanel extends JPanel {
         doRefactorButton.setEnabled(false);
         selectAllButton.setEnabled(false);
         table.setEnabled(false);
-        final List<Refactoring> selectedRefactorings = model.pullSelected().stream().map(CalculatedRefactoring::getRefactoring).collect(Collectors.toList());
-        final List<Refactoring> rejectedRefactorings = refactorings.stream().map(CalculatedRefactoring::getRefactoring).collect(Collectors.toList());
+        final List<MoveToClassRefactoring> selectedRefactorings = model.pullSelected().stream().map(CalculatedRefactoring::getRefactoring).collect(Collectors.toList());
+        final List<MoveToClassRefactoring> rejectedRefactorings = refactorings.stream().map(CalculatedRefactoring::getRefactoring).collect(Collectors.toList());
         rejectedRefactorings.removeAll(selectedRefactorings);
 
         /*
